+++ xoeuf/tests/test_enumeration/tests/test_enumeration.py
@@ -6,21 +6,22 @@
 #
 # This is free software; you can do what the LICENCE file allows you to.
 #
-
 from __future__ import (division as _py3_division,
                         print_function as _py3_print,
                         absolute_import as _py3_abs_import)
 
 import contextlib
+from unittest import skipIf
 
-from hypothesis import strategies as s, given, settings
-from xoeuf import fields
+from hypothesis import strategies as s, given
+
+from xoeuf import fields, MAJOR_ODOO_VERSION
 from xoeuf.odoo.tests.common import TransactionCase, at_install, post_install
 
 from ..models import COLORS, Pax, CARS, WORK_TYPE
 
 colors = s.sampled_from(list(COLORS.__members__.values()))
-color_names = s.sampled_from(list(COLORS.__members__.keys()))
+color_pairs = s.sampled_from(list(COLORS.__members__.items()))
 cars = s.sampled_from(list(CARS.__members__.values()))
 paxs = s.sampled_from(list(Pax.__members__.values()))
 wtypes = s.sampled_from(list(WORK_TYPE.__members__.values()))
@@ -32,10 +33,11 @@ class TestEnum(TransactionCase):
     def setUp(self):
         super(TestEnum, self).setUp()
         self.EnumModel = self.env['test.enum.model']
+        self.DelegatedModel = self.env['test.enum.model_delegated']
 
     def test_column_type_force_char_columns(self):
         self.assertIsInstance(self.EnumModel._fields['color'], fields.Char)
-        self.assertIsInstance(self.EnumModel._fields['color2'], fields.Char)
+        self.assertIsInstance(self.EnumModel._fields['color_name'], fields.Selection)
 
     def test_can_set_valid_integers(self):
         with force_ready(self.env.registry):
@@ -118,69 +120,110 @@ class TestEnum(TransactionCase):
             with self.assertRaises(ValueError):
                 self.EnumModel.search([('color', '=', 10)])
 
-    def test_color2_computed_field_read(self):
-        with force_ready(self.env.registry):
-            obj = self.EnumModel.create({'color2': COLORS.Red})
-            self.assertEqual(obj.color2_name, 'Red')
-
-    def test_color3_default_value(self):
+    def test_color_default_value(self):
         with force_ready(self.env.registry):
             obj = self.EnumModel.create({})
-            self.assertEqual(obj.color3, COLORS.Red)
-
-    @given(color_names)
-    @settings(deadline=None)
-    def test_color2_computed_field_set_on_create(self, name):
-        with force_ready(self.env.registry):
-            obj = self.EnumModel.create({'color2_name': name})
-            self.assertEqual(obj.color2, COLORS.__members__[name])
-
-    @given(color_names)
-    @settings(deadline=None)
-    def test_color2_computed_field_set_on_assignment(self, name):
-        with force_ready(self.env.registry):
-            obj = self.EnumModel.create({})
-            obj.color2_name = name
-            self.assertEqual(obj.color2, COLORS.__members__[name])
-
-    @given(color_names)
-    def test_color2_computed_field_set_on_write(self, name):
-        with force_ready(self.env.registry):
-            obj = self.EnumModel.create({})
-            obj.write({'color2_name': name})
-            self.assertEqual(obj.color2, COLORS.__members__[name])
+            self.assertEqual(obj.color, COLORS.Red)
 
     def test_color_computed_field_read(self):
         with force_ready(self.env.registry):
             obj = self.EnumModel.create({'color': COLORS.Red})
             self.assertEqual(obj.color_name, 'Red')
 
-    @given(color_names)
-    def test_color_computed_field_set_on_create(self, name):
+    @given(color_pairs, color_pairs)
+    def test_color_computed_field_set_on_create(self, pair, update):
         with force_ready(self.env.registry):
+            name, value = pair
             obj = self.EnumModel.create({'color_name': name})
-            self.assertEqual(obj.color, COLORS.__members__[name])
-
-    @given(color_names)
-    def test_color_computed_field_set_on_assignment(self, name):
+            self.assertEqual(obj.color, value)
+            name, value = update
+            obj = self.EnumModel.create({'color_name': name})
+            self.assertEqual(obj.color, value)
         with force_ready(self.env.registry):
+            name, value = pair
+            obj = self.EnumModel.create({'color': value})
+            self.assertEqual(obj.color_name, name)
+            name, value = update
+            obj = self.EnumModel.create({'color': value})
+            self.assertEqual(obj.color_name, name)
+
+    @given(color_pairs, color_pairs)
+    def test_color_computed_field_set_on_assignment(self, pair, update):
+        with force_ready(self.env.registry):
+            name, value = pair
             obj = self.EnumModel.create({})
             obj.color_name = name
-            self.assertEqual(obj.color, COLORS.__members__[name])
-
-    @given(color_names)
-    def test_color_computed_field_set_on_write(self, name):
+            self.assertEqual(obj.color, value)
+            name, value = update
+            obj.color_name = name
+            self.assertEqual(obj.color, value)
         with force_ready(self.env.registry):
+            name, value = pair
+            obj = self.EnumModel.create({})
+            obj.color = value
+            self.assertEqual(obj.color_name, name)
+            name, value = update
+            obj.color = name
+            self.assertEqual(obj.color_name, name)
+
+    @given(color_pairs, color_pairs)
+    def test_color_computed_field_set_on_write(self, pair, update):
+        with force_ready(self.env.registry):
+            name, value = pair
             obj = self.EnumModel.create({})
             obj.write({'color_name': name})
-            self.assertEqual(obj.color, COLORS.__members__[name])
+            self.assertEqual(obj.color, value)
+            name, value = update
+            obj = self.EnumModel.create({})
+            obj.write({'color_name': name})
+            self.assertEqual(obj.color, value)
+        with force_ready(self.env.registry):
+            name, value = pair
+            obj = self.EnumModel.create({})
+            obj.write({'color': value})
+            self.assertEqual(obj.color_name, name)
+            name, value = update
+            obj = self.EnumModel.create({})
+            obj.write({'color': value})
+            self.assertEqual(obj.color_name, name)
 
-    @given(wtypes)
-    def test_intsubclasses_values(self, work_type):
-        from xoeuf.fields.enumeration import _get_db_value
-        value = _get_db_value(self.EnumModel._fields['wtype'], work_type)
-        self.assertIn(value, WORK_TYPE.__members__.keys())
-        self.assertIs(work_type, WORK_TYPE.__members__[value])
+    @skipIf(MAJOR_ODOO_VERSION < 12, "Supported only for  Odoo 12+")
+    @given(color_pairs, color_pairs)
+    def test_color_computed_field_set_on_create_delegated(self, pair, update):
+        with force_ready(self.env.registry):
+            name, value = pair
+            obj = self.DelegatedModel.create({'color_name': name})
+            self.assertEqual(obj.color, value)
+            name, value = update
+            obj = self.DelegatedModel.create({'color_name': name})
+            self.assertEqual(obj.color, value)
+        with force_ready(self.env.registry):
+            name, value = pair
+            obj = self.DelegatedModel.create({'color': value})
+            self.assertEqual(obj.color_name, name)
+            name, value = update
+            obj = self.DelegatedModel.create({'color': value})
+            self.assertEqual(obj.color_name, name)
+
+    @skipIf(MAJOR_ODOO_VERSION < 12, "Supported only for  Odoo 12+")
+    @given(color_pairs, color_pairs)
+    def test_color_computed_field_set_on_assignment_delegated(self, pair, update):
+        with force_ready(self.env.registry):
+            name, value = pair
+            obj = self.DelegatedModel.create({})
+            obj.color_name = name
+            self.assertEqual(obj.color, value)
+            name, value = update
+            obj.color_name = name
+            self.assertEqual(obj.color, value)
+        with force_ready(self.env.registry):
+            name, value = pair
+            obj = self.DelegatedModel.create({})
+            obj.color = value
+            self.assertEqual(obj.color_name, name)
+            name, value = update
+            obj.color = name
+            self.assertEqual(obj.color_name, name)
 
 
 @contextlib.contextmanager